v1
Hello World!

This is everything I've done before moving this repository to GitHub.

This Pokedex was made in early 2013 with around a month of solid work,
followed by over three years of miscellaneous maintenance, upkeep, and
feature addition, most notably the PS teambuilder search feature which
PSdex shares a lot of code with.

I honestly did use Git from the start! But with somewhat lazy commit
messages and entangled with a bunch of other website code... it could
be disentangled, but that'd be a lot of work and I don't think its
historical value is worth that effort, so here we are, restarting from
zero Git history.

Of possible external interest is PS's website's Panel library, with an
open-source license for the first time! Honestly, I've wanted people
to use it forever, but I've never really published it due to worry
about lack of interest. I might try to draw attention to it at some
point later, but you now have my official blessing to use it.

The current license is AGPLv3, same as Pokemon-Showdown-Client. As
with Client, contributors should license their contributions as MIT
License, for ease of relicensing. I haven't decided on a final
license yet... I've been using more copyleft licenses for client
code, but honestly a permissive license might be enough... I don't
think the open-source community will suffer too much if someone
wants to make a proprietary fork of this.

Since this code has, until now, been private to only five developers,
of whom I'm the only one who's ever worked on the code, it's not
particularly easy to jump into right now.

Next steps:
- Add a README
- Make it easier for developers to test changes to contribute
  (something like testclient.html from Pokemon-Showdown-Client?)