Add build script
This is the equivalent of the client's ./githooks/update function
(which should really be renamed). This one is just named ./build and
updates the URL hashes, for easier updates to the dex.

(You'd think there'd be an easier way to do cache invalidation, but
sadly, I haven't found it. If you have, please tell me!)