Add support for articles
On the surface, this is kinda boring, but there's actually a lot of
magic going on.

The build script finds articles in `articles/`, converts them from
pseudo-Markdown and adds them to the dex.

The coolest feature is the `[[Haze]]` wiki-link-style links, which will
automatically figure out that Haze is a move and link to its PSdex move
page.

PSdex is now a pseudo-wiki! :o