Maximize when refocusing right panel
This is mainly to give an easy way to maximize a panel. Previously,
clicking on a link to a panel that already exists (such as the panel
title) scrolls it into view, or, if it's already in view, scrolls it to
the rightmost position so its URL is in the URL bar. Now, if it's
already in the rightmost position, it maximizes it.

(I wonder if we should just always maximize it if it's already in view)